Unbreak Special:InviteSignup
Add missing parameters to UserGroupMembership::getGroupMemberName.
When these were upated, it was not noticed that the replacement
has a mandatory second parameter. When we do not know the gender
of the user, just pass '#' as placeholder that should result in
the unknown gender option.

Also fix a bug where the code was reading 'userid' field, but that
field does not exist. Updated to use the correct field 'invitee'.

Bug: T229349
Change-Id: I32c19101bf8f553cdda512711ff7a6f4136ee679