Update version number because of the important bug fixes
Experimenting with semantic versioning. It is not clear when the
number should be updated (aka when to do a release), and what
counts as a major change (maybe if minimum version of MW is updated?).

I kept the data in the build metadata part as part of the version
string, since the release date is very useful information.

Change-Id: Ic26c3c5ace32a5e0667e8a698a2c90abfdda2b37